adding telemetry to client
